requirements.txt
transformers>=4.36.0
datasets>=2.14.0
evaluate>=0.4.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
flask>=2.3.0
tqdm
